Allow passing datasource, username and password directly to config bean
Bypass a datasource.xml file by passing the datasource name, username and password directly as part of the config bean.  There are suggestions in the code that this is already supported but it was missing the actual implementation.  The config bean already supports these properties, the init just needs to support setting them so it can be passed in a coldspring declaration like so:

```
<bean id="transferConfiguration" class="transfer.com.config.Configuration">
	<constructor-arg name="name"><value>myDSN</value></constructor-arg>
	<constructor-arg name="username"><value>myUsername</value></constructor-arg>
	<constructor-arg name="password"><value>myPassword</value></constructor-arg>
	<constructor-arg name="configPath"><value>/config/transfer/transfer.xml</value></constructor-arg>
	<constructor-arg name="definitionPath"><value>/tmp/generated</value></constructor-arg>
</bean>
```

The reason to make this change was to pass in the datasource dynamically from an environment manager:

```
<bean id="transferConfiguration" class="transfer.com.config.Configuration">
	<constructor-arg name="name"><value><bean factory-bean="globalConfig" factory-method="get"><constructor-arg name="name"><value>DATASOURCE</value></constructor-arg></bean></value></constructor-arg>
	<constructor-arg name="configPath"><value>/config/transfer/transfer.xml</value></constructor-arg>
	<constructor-arg name="definitionPath"><value>/tmp/generated</value></constructor-arg>
</bean>
```

This eliminates the need for a datasource.xml, similar to this Coldbox add-on TransferConfigFactory.cfc: http://wiki.coldbox.org/wiki/Extras:TransferORM.cfm